fix(record): split peekUpstreamLive by GOOS so Windows builds
The MSG_PEEK probe added in this PR uses syscall.Recvfrom and the
MSG_PEEK / MSG_DONTWAIT constants, which Go's standard syscall package
exposes only on POSIX. Windows fails to build (#4166 CI):

  pkg\agent\proxy\incoming\http.go:944:34: cannot use int(fd) as
    syscall.Handle in argument to syscall.Recvfrom
  pkg\agent\proxy\incoming\http.go:944:59: undefined: syscall.MSG_PEEK
  pkg\agent\proxy\incoming\http.go:944:76: undefined: syscall.MSG_DONTWAIT

Move the function out of http.go into platform-specific files:
- peek_unix.go: keeps the existing recvfrom(MSG_PEEK|MSG_DONTWAIT) impl
- peek_windows.go: equivalent via golang.org/x/sys/windows.WSARecv with
  windows.MSG_PEEK; MSG_DONTWAIT has no analogue because Go puts the
  socket in non-blocking IOCP mode already, so an empty queue returns
  WSAEWOULDBLOCK immediately.
